Move extractSentence
