+++ ext/js/dom/document-util.js
@@ -16,142 +16,12 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {isWhitespace} from '../data/sandbox/string-util.js';
-
 /**
  * This class contains utility functions related to the HTML document.
  */
 export class DocumentUtil {
     /** @type {?boolean} */
     static _cssZoomSupported = null;
-
-    /**
-     * Extract a sentence from a document.
-     * @param {import('text-source').TextSource} source The text source object, either `TextSourceRange` or `TextSourceElement`.
-     * @param {boolean} layoutAwareScan Whether or not layout-aware scan mode should be used.
-     * @param {number} extent The length of the sentence to extract.
-     * @param {boolean} terminateAtNewlines Whether or not a sentence should be terminated at newline characters.
-     * @param {import('text-scanner').SentenceTerminatorMap} terminatorMap A mapping of characters that terminate a sentence.
-     * @param {import('text-scanner').SentenceForwardQuoteMap} forwardQuoteMap A mapping of quote characters that delimit a sentence.
-     * @param {import('text-scanner').SentenceBackwardQuoteMap} backwardQuoteMap A mapping of quote characters that delimit a sentence, which is the inverse of forwardQuoteMap.
-     * @returns {{text: string, offset: number}} The sentence and the offset to the original source.
-     */
-    static extractSentence(source, layoutAwareScan, extent, terminateAtNewlines, terminatorMap, forwardQuoteMap, backwardQuoteMap) {
-        // Scan text
-        source = source.clone();
-        const startLength = source.setStartOffset(extent, layoutAwareScan);
-        const endLength = source.setEndOffset(extent * 2 - startLength, true, layoutAwareScan);
-        const text = source.text();
-        const textLength = text.length;
-        const textEndAnchor = textLength - endLength;
-
-        /** Relative start position of the sentence (inclusive). */
-        let cursorStart = startLength;
-        /** Relative end position of the sentence (exclusive). */
-        let cursorEnd = textEndAnchor;
-
-        // Move backward
-        let quoteStack = [];
-        for (; cursorStart > 0; --cursorStart) {
-            // Check if the previous character should be included.
-            let c = text[cursorStart - 1];
-            if (c === '\n' && terminateAtNewlines) { break; }
-
-            if (quoteStack.length === 0) {
-                let terminatorInfo = terminatorMap.get(c);
-                if (typeof terminatorInfo !== 'undefined') {
-                    // Include the previous character while it is a terminator character and is included at start.
-                    while (terminatorInfo[0] && cursorStart > 0) {
-                        --cursorStart;
-                        if (cursorStart === 0) { break; }
-                        c = text[cursorStart - 1];
-                        terminatorInfo = terminatorMap.get(c);
-                        if (typeof terminatorInfo === 'undefined') { break; }
-                    }
-                    break;
-                }
-            }
-
-            let quoteInfo = forwardQuoteMap.get(c);
-            if (typeof quoteInfo !== 'undefined') {
-                if (quoteStack.length === 0) {
-                    // Include the previous character while it is a quote character and is included at start.
-                    while (quoteInfo[1] && cursorStart > 0) {
-                        --cursorStart;
-                        if (cursorStart === 0) { break; }
-                        c = text[cursorStart - 1];
-                        quoteInfo = forwardQuoteMap.get(c);
-                        if (typeof quoteInfo === 'undefined') { break; }
-                    }
-                    break;
-                } else if (quoteStack[0] === c) {
-                    quoteStack.pop();
-                    continue;
-                }
-            }
-
-            quoteInfo = backwardQuoteMap.get(c);
-            if (typeof quoteInfo !== 'undefined') {
-                quoteStack.unshift(quoteInfo[0]);
-            }
-        }
-
-        // Move forward
-        quoteStack = [];
-        for (; cursorEnd < textLength; ++cursorEnd) {
-            // Check if the following character should be included.
-            let c = text[cursorEnd];
-            if (c === '\n' && terminateAtNewlines) { break; }
-
-            if (quoteStack.length === 0) {
-                let terminatorInfo = terminatorMap.get(c);
-                if (typeof terminatorInfo !== 'undefined') {
-                    // Include the following character while it is a terminator character and is included at end.
-                    while (terminatorInfo[1] && cursorEnd < textLength) {
-                        ++cursorEnd;
-                        if (cursorEnd === textLength) { break; }
-                        c = text[cursorEnd];
-                        terminatorInfo = terminatorMap.get(c);
-                        if (typeof terminatorInfo === 'undefined') { break; }
-                    }
-                    break;
-                }
-            }
-
-            let quoteInfo = backwardQuoteMap.get(c);
-            if (typeof quoteInfo !== 'undefined') {
-                if (quoteStack.length === 0) {
-                    // Include the following character while it is a quote character and is included at end.
-                    while (quoteInfo[1] && cursorEnd < textLength) {
-                        ++cursorEnd;
-                        if (cursorEnd === textLength) { break; }
-                        c = text[cursorEnd];
-                        quoteInfo = forwardQuoteMap.get(c);
-                        if (typeof quoteInfo === 'undefined') { break; }
-                    }
-                    break;
-                } else if (quoteStack[0] === c) {
-                    quoteStack.pop();
-                    continue;
-                }
-            }
-
-            quoteInfo = forwardQuoteMap.get(c);
-            if (typeof quoteInfo !== 'undefined') {
-                quoteStack.unshift(quoteInfo[0]);
-            }
-        }
-
-        // Trim whitespace
-        for (; cursorStart < startLength && isWhitespace(text[cursorStart]); ++cursorStart) { /* NOP */ }
-        for (; cursorEnd > textEndAnchor && isWhitespace(text[cursorEnd - 1]); --cursorEnd) { /* NOP */ }
-
-        // Result
-        return {
-            text: text.substring(cursorStart, cursorEnd),
-            offset: startLength - cursorStart
-        };
-    }
 
     /**
      * Computes the scaling adjustment that is necessary for client space coordinates based on the
